Extrair sub-componentes para reduzir nesting de divs
Novos componentes (depth 1-2 cada):
- HeatmapYearNav: navegação de ano (← 2024 →, Hoje)
- HeatmapMonthPicker: grid de 12 meses com barras de progresso
- HeatmapLegend: legenda com swatches (usa <ul> em vez de div>div>div)
- CoverageTable: tabela ordenável com lógica de sort encapsulada
- TribunalCard: card por tribunal no grid (link + article)
- TribunalStatsBar: barra de 3 stats usando <dl> semântico

Pais refatorados:
- Heatmap.svelte: depth 6 → 2 (extrai 3 sub-componentes)
- AnnualCoverageMonitor.svelte: delega YearSummaryCards + CoverageTable
- TribunalView.svelte: loop usa TribunalCard
- TribunalDetail.svelte: stats usa TribunalStatsBar

TribunalDetail/View/AnnualCoverageMonitor ainda têm nesting residual
nos seus próprios wrappers — segunda rodada de extração necessária.
