+++ web/src/components/HeatmapLegend.svelte
@@ -1,33 +1,24 @@
 <script lang="ts">
-  interface HeatmapLegendProps {
-    coveredDays: number;
-    totalDays: number;
-  }
-
-  let { coveredDays, totalDays }: HeatmapLegendProps = $props();
+  let { coveredDays, totalDays }: { coveredDays: number; totalDays: number } = $props();
 </script>
 
-<div class="legend-card">
-  <div class="legend-body">
-    <span class="legend-summary">
-      <strong>{coveredDays}</strong> de <strong>{totalDays}</strong> dias com dados neste mês
-    </span>
-    <div class="legend-items">
-      <span class="legend-label">Legenda:</span>
-      <div class="legend-item"><div class="legend-swatch heatmap-missing"></div><span>Faltante</span></div>
-      <div class="legend-item"><div class="legend-swatch heatmap-absent"></div><span>Ausente</span></div>
-      <div class="legend-item"><div class="legend-swatch heatmap-collected"></div><span>Coletado</span></div>
-    </div>
-  </div>
+<div class="legend-strip">
+  <span class="legend-summary">
+    <strong>{coveredDays}</strong> de <strong>{totalDays}</strong> dias com dados neste mês
+  </span>
+  <ul class="legend-swatches" aria-label="Legenda">
+    <li class="heatmap-missing">Faltante</li>
+    <li class="heatmap-absent">Ausente</li>
+    <li class="heatmap-collected">Coletado</li>
+  </ul>
 </div>
 
 <style>
-  .legend-card   { background: var(--color-base-200); border-radius: var(--radius-box); }
-
-  .legend-body {
+  .legend-strip {
+    background: var(--color-base-200);
+    border-radius: var(--radius-box);
     padding: 0.75rem 1rem;
     display: flex;
-    flex-direction: row;
     justify-content: space-between;
     align-items: center;
     font-size: var(--font-size-sm);
@@ -37,20 +28,29 @@
 
   .legend-summary { opacity: 0.7; }
 
-  .legend-items {
+  .legend-swatches {
     display: flex;
     align-items: center;
     flex-wrap: wrap;
-    gap: 1rem;
+    gap: 0.75rem;
+    list-style: none;
+    margin: 0;
+    padding: 0;
   }
 
-  .legend-label { opacity: 0.5; }
+  .legend-swatches li {
+    display: flex;
+    align-items: center;
+    gap: 0.375rem;
+    font-size: var(--font-size-xs);
+  }
 
-  .legend-item  { display: flex; align-items: center; gap: 0.375rem; }
-
-  .legend-swatch {
+  .legend-swatches li::before {
+    content: '';
+    display: inline-block;
     width: 0.75rem;
     height: 0.75rem;
     border-radius: var(--radius-sm);
+    background: currentColor;
   }
 </style>
